Fixed Partent/Child mess at Container Linking
First the docker ps command output showing the two linked containers is wrong. When running the commands, the outcome is different. I fixed that. Further there seems to be a misunderstanding about who is parent and who is the child.

For me it's pretty clear that web is the parent and db is the child. Because the web containers knows about the db container and can access it. Also you can use more than one --link parameter in one command in order to link multiple childs to the parent.